+++ src/main/java/net/silentchaos512/gems/setup/GemsTags.java
@@ -5,6 +5,7 @@ import net.minecraft.resources.ResourceLocation;
 import net.minecraft.tags.BlockTags;
 import net.minecraft.tags.ItemTags;
 import net.minecraft.tags.TagKey;
+import net.minecraft.world.damagesource.DamageType;
 import net.minecraft.world.entity.EntityType;
 import net.minecraft.world.item.Item;
 import net.minecraft.world.level.block.Block;
@@ -66,6 +67,11 @@ public class GemsTags {
 
     public static final class EntityTypes {
         public static final TagKey<EntityType<?>> COFFEE_PRODUCERS = gems("coffee_producers");
+        public static final TagKey<EntityType<?>> END_MONSTERS = gems("end_monsters");
+
+        private static TagKey<EntityType<?>> common(String path) {
+            return tag("c", path);
+        }
 
         private static TagKey<EntityType<?>> gems(String path) {
             return tag(SilentGems.MOD_ID, path);
@@ -73,6 +79,22 @@ public class GemsTags {
 
         private static TagKey<EntityType<?>> tag(String namespace, String path) {
             return TagKey.create(Registries.ENTITY_TYPE, ResourceLocation.fromNamespaceAndPath(namespace, path));
+        }
+    }
+
+    public static final class DamageTypes {
+        public static final TagKey<DamageType> NEPTUNES_BLESSING_PROTECTS = gems("neptunes_blessing_protects");
+
+        private static TagKey<DamageType> common(String path) {
+            return tag("c", path);
+        }
+
+        private static TagKey<DamageType> gems(String path) {
+            return tag(SilentGems.MOD_ID, path);
+        }
+
+        private static TagKey<DamageType> tag(String namespace, String path) {
+            return TagKey.create(Registries.DAMAGE_TYPE, ResourceLocation.fromNamespaceAndPath(namespace, path));
         }
     }
 }