+++ CHANGELOG.md
@@ -3,9 +3,12 @@
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [1.21.1-5.1.2] - 2025-07-13
+Requires Silent Gear 4.0.25 or higher
 ### Added
 - Five new gems: garnet, aquamarine, tanzanite, opal, and pearl
+- New traits
+- Reinforced gold/silver rods
 ### Fixed
 - Redstone teleporters will no longer move entities in the `c:teleporting_not_supported` tag and will not move entities across dimensions if the entity normally does not allow it.